ShowVideoNameOverlay v1.0.0
Always shows user's name overlay when viewing video call cameras. Does not work in video pop out.

Also changes the mic activity circle to a red color (was using for testing but left it for personal preference.) Find and replace "#BA3B46" with your own hex color to customize it!